🧹 Get resources route working
This commit will get the resources route working. It will also add a
HykuKnapsack::PagesController which inherits from Hyrax::PagesController
so the route won't throw a namespace error.  Further down the line, we
needed to decorate the ContentBlock model to add the resources key.  I
am doing this with no change to Hyku although we may want to add the
`#name_registry` method to Hyku as a convenience for overrides since we
can't add to a frozen hash.