🧹 Remove spec for NavigationPresenter
Prior to this commit, it appears that knapsack's spec version was a copy
of a prior Hyku state's spec version.  When I ran the now deleted
Knapsack spec, it failed.  I copied over the updated Hyku spec ran it
and all things passed.

As such, the NavigationPresenter spec appears to be a duplicate and can
go away.

Related:

- https://github.com/scientist-softserv/adventist-dl/issues/538