🧹 Handle slugs
Fixes https://github.com/scientist-softserv/adventist-dl/issues/582

Removes `alternative_title` from forms
Removes `alternative_title` from bulkrax mappings

Additional Background:
In the adventist implementation of slugs, the predicate "alternative"
is being used to store the slug. If viewing the data, the predicate will
contain the slug if there is one, and otherwise it contains the uuid.

A second term and predicate was added for slugs, but remediation of
prior data into the new predicate has not yet happened, which
would be beyond the scope of current work as it would require a data
migration.

With the hyrax upgrade, alternative_title is now included by default.
By removing the alternative_title from the any create or edit ability,
the data for slugs is still saved in the same location as alternative_title would be, but does not impede any behavior.