🐛 Ensure we pass query to IIIF viewer
In the pre-knapsack Adventist, the IIIF Print gem overwrote :
`app/views/catalog/_index_header_list_default.html.erb` and the
Adventist Hyku instance did not have that view.

However with the upgrade to the latest Hyku version via Knapsack, that
Hyku overrided Hyrax's view.  And due to view_path issues, we favor the
Hyku instance.

The newer Hyku view override leverages a method that we in the
knapsack (and later in IIIF Print) can extend to inject the parent_query
and query parameter necessary for OCR highlighting.

Below is a *diff of the files in Knapsack's Hyku and IIIF Print*.  I
have not resolved all of these things but am close:

Note that when 'q' is present AND `document.any_highlighting?` is true
we want to pass the param as "parent_query".  If just 'q' is present, we
want to pass "query".  So this is only a partial fix.

<details>
<summary>diff of the files in Knapsack's Hyku and IIIF Print</summary>

```
❯ diff hyrax-webapp/app/views/catalog/_index_header_list_default.html.erb ../iiif_print/app/views/catalog/_index_header_list_default.html.erb
1,2c1,2
< <%# OVERRIDE Hyrax 3.4.0 to support shared search %>
< <% model = document.hydra_model %>
---
> <%# OVERRIDE Hyrax 2.9.6 to show parent_query params if metadata is found in parent record %>
>
4,9c4,12
<     <% if model == Hyrax::PcdmCollection || model < Hyrax::PcdmCollection %>
<         <h4 class="search-result-title"><%= link_to document.title_or_label, generate_work_url(document, request) %></h4>
<         <%= Hyrax::CollectionPresenter.new(document, current_ability).collection_type_badge %>
<     <% else %>
<         <h4 class="search-result-title"><%= link_to document.title_or_label, generate_work_url(document, request) %></h4>
<     <% end %>
---
>   <h3 class="search-result-title">
>   <% if params['q'].present? && document.any_highlighting? %>
>     <%= link_to document.title_or_label, [document, { parent_query: params['q'] }] %></h3>
>   <% elsif params['q'].present? %>
>     <%= link_to document.title_or_label, [document, { query: params['q'] }] %></h3>
>   <% else %>
>     <%= link_to document.title_or_label, document %></h3>
>   <% end %>
>   </h3>
```

</details>

Co-authored-by: Kirk Wang <kirk.wang@scientist.com>