🧹 Comment out link color selection assignment
Assigning the link color to be used in the footer was overriding it to
the link_color in Hyku rather than the one from this app's
appearance_styles view. It loaded the appearance style correctly but at
some point, the value got change to the one in Hyku, resulting in an
incorrect footer text link color.

Commenting this out did not seem to result in any other links to be
incorrect color, but this should be corrected in future UI work.