🐛 Render publication_status as multi-value
Prior to this commit, we specified that publication_status was a
singular value.  However that broke the general paradigm that most all
Hyrax model properties are multi-value.

What happened was that we would submit `collection[publication_status]`
which was ignored.  When we submit `collection[publication_status][]`
then the property value persists.

We also needed to consider how we'd translate the drop-down value to a
human readable value.  That is done in the corresponding show_fields
page.

Related to:

- https://github.com/scientist-softserv/adventist-dl/issues/620