🐛 Fix add to collection by removing spellcheck
In looking at the responses, Solr was complaining about a missing
spellcheck.dictionary for contributor.

To get this working, we removed the `spellcheck.dictionary` declaration.
One observation we had is that matches are now only made when we type a
significant number of characters (or 2 words, we don't know what).

An observed ongoing issue is that we're only finding exact matches.

Related:

- https://github.com/scientist-softserv/adventist-dl/issues/623
- https://github.com/scientist-softserv/adventist-dl/issues/622

Co-authored-by: LaRita Robinson <larita@scientist.com>