🐛 Make date created consistent
Ref:
- https://github.com/scientist-softserv/adventist-dl/issues/632
Related to:
- https://github.com/scientist-softserv/adventist_knapsack/pull/77

Property date_created was defined inconsistently, which broke creation
and editing of Image and GenericWork types.

This pull request adds a partial with a date picker for date_created and
removes `multiple: false` from the definition for GenericWork and
Image work types.

Additionally, several unused view partials are being removed, as they
were in the wrong path location and would never be hit.