🐛 Return identifier to collection show
The identifier was removed from collections via [PR 16][1].  And in #620
we have been asked to restore that.

I have looked and the collection does not have an Alternate Title field,
so I don't know how to address/resolve the request for that property to
show.

There are complications with the alternative (as envisioned in Hyrax)
because we have overloaded the meaning of that predicate to be the slug.
This will create data issues if we choose to re-introduce alternative
title.

Related to:

- https://github.com/scientist-softserv/adventist-dl/issues/620
- https://github.com/scientist-softserv/adventist-dl/issues/578

[1]: https://github.com/scientist-softserv/adventist_knapsack/pull/16