♻️ Move decoration to after initialize
This is a guess as to some of the loading issues.  Namely that I have
specs that say the code for a class should be in "one file" but is in a
"different one".  That indicates the "different one" is being loaded
after the "one file".

So, I'm looking to see if I can adjust the load sequence of the
decorators.

Update, based on one run with these specs in the test suite, things
worked.  Hopefully this alleviates further problems.