Add a listener to remove solr for slugs
After deleting a work, Hyrax's listener removes the index.

`Hyrax.index_adapter.delete(resource: event[:object])`

In the case of slugs, this wouldn't find the index since solr's id
is the slug, not the UUID of the resource. This commit overrides Hyrax's
listener to remove the indexes by query instead.