💄 Refine css for adventist_digital_library theme
This commit will move the css for the adventist_digital_library theme to
only affect things when it is on that theme.  Also removed the
home_content partial since it was not used in the theme.  It would throw
and error if used on the Default Theme so it was no longer needed.  Also
added a screenshot for the adventist_digital_library theme so it can be
previewed in the admin appearance themes section.