🐛 Get catalog index working
There was an override in the knapsack that wasn't needed anymore and it
was causing the catalog index page to crash.  Also, added the themes
behavior mixin to the catalog controller as well because of the custom
user_util_links and controls partial that is used.  We need to inject
those paths into the view paths.