✅ Fixes failing specs
This commit fixes the failing application_spec and ignores the splash
page spec that is currently failing due to the changes in styling. The
client requested the styling be reverted to base Hyku. We are waiting to
get feedback from the client on the styling of the splash page. If the
page stays as is, this spec can be removed. If the styling is updated,
this spec will need to be updated.

Ref:
- [741](https://github.com/scientist-softserv/adventist_knapsack/issues/741)