🐛 Wrap slug_valkyrie_overrides in to_prepare
Originally we had a require 'frigg' in this initializer because we want
frigg to log and then load the slug logic.  However, this threw off the
load order of things and caused the Hyrax::PcdmCollection to be
registered for Collection instead of CollectionResource.  Wrapping this
in a to_prepare will ensure it's loaded at the right time and keep it
loaded in dev.