🚧 WIP Prevents rendering blank metadata fields
This commit modifies the metadata yaml files for collections and all
work types removing the 'm' from solr index keys for values that
contain `multiple: false`.

This fixes the issue of rendering blank metadata fields for all items
except author (creator). It does not fix the underlying issue of fields
being assigned `[""]` when they should be `[]`.

Ref:
- https://github.com/scientist-softserv/adventist_knapsack/issues/771